Feat/audio buffer source node/override start method (#297)
* feat: implemented overridden buffer source start method

* feat: added pause feature to example app

* chore: update API coverage

---------

+++ packages/react-native-audio-api/common/cpp/core/AudioBuffer.cpp
@@ -27,8 +27,8 @@ float AudioBuffer::getSampleRate() const {
   return bus_->getSampleRate();
 }
 
-float AudioBuffer::getDuration() const {
-  return static_cast<float>(getLength()) / getSampleRate();
+double AudioBuffer::getDuration() const {
+  return static_cast<double>(getLength()) / getSampleRate();
 }
 
 float *AudioBuffer::getChannelData(int channel) const {
